requirements.txt
ipdb
timm
Cython
opencv-python
imageio
pyviz3d
natsort
plyfile
prettytable
torchinfo
pyyaml==5.3.1
scipy
transformers
ipdb
Pillow==9.1.0
wandb
einops
pandas
scikit-learn
ai2thor
prior
shapely
fire
loguru
git+https://github.com/facebookresearch/detectron2.git
git+https://github.com/facebookresearch/pytorch3d.git@stable
git+https://github.com/waspinator/pycococreator.git@0.2.0
git+https://github.com/waspinator/coco.git@2.1.0

